Refactor entities, models & types
