+++ src/index.ts
@@ -1,78 +1,8 @@
-export interface Address {
-  street: string;
-  city: string;
-  country: string;
-  postalCode: number;
-}
+import { Employee } from "./entities/employee";
+import debug from "debug";
 
-
-// TODO: La clase Person tiene los siguientes atributos: id, firstName, lastName y address. 👈
-export class Person {
-  constructor(
-    protected id: number, // TODO: El id debería permitir almacenar tanto números como cadenas. 👈
-    protected firstName: string,
-    protected lastName: string,
-    protected address: string // TODO: El address debería ser de usar la interfaz Address 👈
-  ) {}
-}
-
-
-// TODO: La clase Employee debería heredar de la clase Person. 👈
-export class Employee {
-  constructor(
-    id: string, // TODO: El id debería permitir almacenar tanto números como cadenas. 👈
-    firstName: string,
-    lastName: string,
-    address: null, // TODO: El address debería ser de usar la interfaz Address 👈
-    private department: string,
-    private paymentPerHour: number,
-    private workingHours: number,
-    private hireDate: string // TODO: Se debería manejar la fecha como tipo Date y debería ser opcional 👈
-  ) {
-    super(id, firstName, lastName, address);
-  }
-
-  showInfo(): string {
-    const fullName = this.displayFullName();
-    const department = `Department: ${this.department}`;
-    const address = this.formatAddress();
-    const hireDate = `Hire Date: ${
-      this.hireDate == null ? "-" : this.hireDate.toLocaleString()
-    }`;
-    const seniority = `Seniority: ${this.calculateSeniority()}`;
-    const salary = `Salary: ${this.calculateSalary()}`;
-    return `
-      Employee ${this.id}:
-      ${fullName}
-      ${department}
-      ${address}
-      ${hireDate}
-      ${seniority}
-      ${salary}
-    `;
-  }
-
-  formatAddress(): string {
-    return `Address: ${this.address.street} ${
-      this.address.postalCode == null ? "" : "#" + this.address.postalCode
-    }, ${this.address.city} - ${this.address.country}`;
-  }
-
-  displayFullName(): string {
-    return `Name: ${this.firstName} ${this.lastName}`;
-  }
-
-  calculateSalary(): number {
-    return this.paymentPerHour * this.workingHours;
-  }
-
-  calculateSeniority(): number {
-    if (this.hireDate) {
-      return new Date().getFullYear() - this.hireDate.getFullYear();
-    }
-    return 0;
-  }
-}
+const ORIGIN_MODULE = 'ts-day:Index';
+const logger = debug(ORIGIN_MODULE);
 
 const newEmployee = new Employee(
   "100",
@@ -84,4 +14,6 @@ const newEmployee = new Employee(
   160,
   new Date("2019-02-01")
 );
-newEmployee.showInfo();
+
+const employeeInfo = newEmployee.showInfo();
+logger('Employe Info %s', employeeInfo);
