+++ src/models/i-addres.ts
@@ -1,0 +1,6 @@
+export interface Address {
+  street: string;
+  city: string;
+  country: string;
+  postalCode: number;
+}
