+++ src/types/person-types.ts
@@ -1,0 +1,1 @@
+export type id = string | number;
